Group_15/backend/pyproject.toml
[project]
name = "ideascope-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.89.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain-anthropic>=1.4.0",
    "langgraph>=1.1.6",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.43.0",
]
